requirements.txt
Flask==2.3.1
Jinja2==2.11.3
Werkzeug==0.15.4
PyYAML==3.13